#27 루트 카테고리 삭제 예외처리 (develop -> master)

+++ linknamu/src/main/java/com/kakao/linknamu/category/service/CategoryDeleteService.java
@@ -1,5 +1,7 @@
 package com.kakao.linknamu.category.service;
 
+import com.kakao.linknamu._core.exception.Exception400;
+import com.kakao.linknamu.category.CategoryExceptionStatus;
 import com.kakao.linknamu.category.entity.Category;
 import com.kakao.linknamu.user.entity.User;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +15,9 @@ public class CategoryDeleteService {
 
     public void delete(Long categoryId, User user) {
         Category category = categoryService.findById(categoryId);
+        if (category.getParentCategory().getCategoryId().equals(categoryId)){
+            throw new Exception400(CategoryExceptionStatus.ROOT_DELETION_FORBIDDEN);
+        }
         categoryService.validUser(category, user);
         categoryService.deleteById(categoryId);
     }
